Messing with NuGet packages.
Ultimately, I plan to have this work the same way that Harmony does: mod developers add the NuGet package/dll, but don't copy to output. They then add a dependency on the mod, which loads the 'master' dll.
Finally then the game loads, the master mod loads the master dll, the mods that add wiki's now have access to the API, and everyone is happy. If the master mod/API updates, the each mod author has to download the latest NuGet package and re-publish their mod to fix any conflicts.